+++ src/main/java/com/example/queue_management_system_hospital/repo/AppointmentRepo.java
@@ -5,5 +5,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AppointmentRepo extends JpaRepository<Appointment,Long> {
-}
+public interface AppointmentRepo extends JpaRepository<Appointment,Integer> {}
